Move UDF generator to algorithms subpackage
