+++ tox.ini
@@ -5,7 +5,7 @@ skipsdist = True
 basepython = python3.8
 deps = -rrequirements/tests.txt
         -rrequirements/node.txt
-
+        -rrequirements/algorithms.txt
 
 [testenv:api]
 changedir = mipengine/tests/controller/api
@@ -14,5 +14,3 @@ commands = python -m pytest -n 5
 [testenv:node]
 changedir = mipengine/tests/node
 commands = python -m pytest
-
-
